Change tests location and implement pre-commit
