+++ cmd/secret_test.go
@@ -1,49 +1,88 @@
-package testsecrets
+package cmd
 
 import (
+	"fmt"
 	"testing"
 
-	"github.com/Bisnode/kubectl-tbac/cmd"
-	"github.com/Bisnode/kubectl-tbac/tests/testdata"
 	"github.com/stretchr/testify/assert"
+	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes/fake"
 )
+
+// GenerateSecrets is a set of secret definitions.
+var GenerateSecrets = []v1.Secret{
+	v1.Secret{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "my-credentials",
+			Namespace: "default",
+			Labels: map[string]string{
+				"app":                        "my-credentials",
+				"tbac.bisnode.com/container": "default",
+			},
+			Annotations: map[string]string{
+				"tbac.bisnode.com/last-modified": fmt.Sprintf("%v", metav1.Now().Rfc3339Copy()),
+			},
+		},
+		Data: map[string][]byte{
+			"USERNAME": []byte("foo"),
+			"PASSWORD": []byte("bar"),
+			"KEY":      []byte("extra-key"),
+		},
+	},
+	v1.Secret{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "my-api-key",
+			Namespace: "default",
+			Labels: map[string]string{
+				"app":                        "my-api-key",
+				"tbac.bisnode.com/container": "default",
+			},
+			Annotations: map[string]string{
+				"tbac.bisnode.com/last-modified": fmt.Sprintf("%v", metav1.Now().Rfc3339Copy()),
+			},
+		},
+		Data: map[string][]byte{
+			"URL": []byte("github.com"),
+			"KEY": []byte("key"),
+		},
+	},
+}
 
 func TestGetSecretsList(t *testing.T) {
 	// create the 'fake' clientSet where clientSet.Interface = &Clientset{}, setting all the 'fake' methods
 	// as seen in https://github.com/kubernetes/client-go/blob/master/kubernetes/fake/clientSet_generated.go
 	clientSet := fake.NewSimpleClientset()
-	cmd.Namespace = "default"
+	Namespace = "default"
 
-	for _, s := range testdata.GenerateSecrets {
-		_, err := clientSet.CoreV1().Secrets(cmd.Namespace).Create(&s)
+	for _, s := range GenerateSecrets {
+		_, err := clientSet.CoreV1().Secrets(Namespace).Create(&s)
 		if err != nil {
 			assert.Equal(t, nil, err)
 		}
 	}
 
-	secretList, err := cmd.GetSecretList(clientSet)
+	secretList, err := GetSecretList(clientSet)
 	if err != nil {
 		assert.Equal(t, nil, err)
 	}
-	assert.Equal(t, len(testdata.GenerateSecrets), len(secretList))
+	assert.Equal(t, len(GenerateSecrets), len(secretList))
 	assert.Contains(t, secretList, "my-credentials")
 	assert.Contains(t, secretList, "my-api-key")
 }
 
 func TestDescribeOneSecret(t *testing.T) {
 	clientSet := fake.NewSimpleClientset()
-	cmd.Namespace = "default"
+	Namespace = "default"
 
-	for _, s := range testdata.GenerateSecrets {
-		_, err := clientSet.CoreV1().Secrets(cmd.Namespace).Create(&s)
+	for _, s := range GenerateSecrets {
+		_, err := clientSet.CoreV1().Secrets(Namespace).Create(&s)
 		if err != nil {
 			assert.Equal(t, nil, err)
 		}
 	}
 
-	secretDescription, err := cmd.GetSecretDescription(clientSet, "my-credentials")
+	secretDescription, err := GetSecretDescription(clientSet, "my-credentials")
 	if err != nil {
 		assert.Equal(t, nil, err)
 	}
@@ -54,20 +93,20 @@ func TestDescribeOneSecret(t *testing.T) {
 
 func TestDeleteSecret(t *testing.T) {
 	clientSet := fake.NewSimpleClientset()
-	cmd.Namespace = "default"
+	Namespace = "default"
 
-	for _, s := range testdata.GenerateSecrets {
-		_, err := clientSet.CoreV1().Secrets(cmd.Namespace).Create(&s)
+	for _, s := range GenerateSecrets {
+		_, err := clientSet.CoreV1().Secrets(Namespace).Create(&s)
 		if err != nil {
 			assert.Equal(t, nil, err)
 		}
 	}
 
-	err := cmd.DeleteSecret(clientSet, "my-credentials")
+	err := DeleteSecret(clientSet, "my-credentials")
 	if err != nil {
 		assert.Equal(t, nil, err)
 	}
-	secretList, err := clientSet.CoreV1().Secrets(cmd.Namespace).List(metav1.ListOptions{})
+	secretList, err := clientSet.CoreV1().Secrets(Namespace).List(metav1.ListOptions{})
 	if err != nil {
 		assert.Equal(t, nil, err)
 	}
@@ -81,10 +120,10 @@ func TestDeleteSecret(t *testing.T) {
 
 func TestPatchSecret(t *testing.T) {
 	clientSet := fake.NewSimpleClientset()
-	cmd.Namespace = "default"
+	Namespace = "default"
 
-	for _, s := range testdata.GenerateSecrets {
-		_, err := clientSet.CoreV1().Secrets(cmd.Namespace).Create(&s)
+	for _, s := range GenerateSecrets {
+		_, err := clientSet.CoreV1().Secrets(Namespace).Create(&s)
 		if err != nil {
 			assert.Equal(t, nil, err)
 		}
@@ -97,12 +136,12 @@ func TestPatchSecret(t *testing.T) {
 		"URL=my-api.com",
 	}
 
-	err := cmd.PatchSecret(clientSet, &secretName, &removeData, &updateData)
+	err := PatchSecret(clientSet, &secretName, &removeData, &updateData)
 	if err != nil {
 		assert.Equal(t, nil, err)
 	}
 
-	updatedSecret, err := clientSet.CoreV1().Secrets(cmd.Namespace).Get("my-credentials", metav1.GetOptions{})
+	updatedSecret, err := clientSet.CoreV1().Secrets(Namespace).Get("my-credentials", metav1.GetOptions{})
 	if err != nil {
 		assert.Equal(t, nil, err)
 	}
@@ -118,7 +157,7 @@ func TestPatchSecret(t *testing.T) {
 
 func TestCreateSecret(t *testing.T) {
 	clientSet := fake.NewSimpleClientset()
-	cmd.Namespace = "default"
+	Namespace = "default"
 
 	secretName := "new-app-secret"
 	container := "default"
@@ -126,11 +165,11 @@ func TestCreateSecret(t *testing.T) {
 		"USERNAME=foo",
 		"PASSWORD=bar",
 	}
-	err := cmd.CreateSecret(clientSet, &secretName, &container, data)
+	err := CreateSecret(clientSet, &secretName, &container, data)
 	if err != nil {
 		assert.Equal(t, nil, err)
 	}
-	createdSecret, err := clientSet.CoreV1().Secrets(cmd.Namespace).Get("new-app-secret-default", metav1.GetOptions{})
+	createdSecret, err := clientSet.CoreV1().Secrets(Namespace).Get("new-app-secret-default", metav1.GetOptions{})
 	if err != nil {
 		assert.Equal(t, nil, err)
 	}
